requirements.txt
streamlit
replicate
openai
pyyaml